seperate counter module
