+++ react/package.json
@@ -16,6 +16,7 @@
     "react-redux": "^7.2.2",
     "react-router": "^5.2.0",
     "redux": "^4.0.5",
+    "redux-saga": "^1.1.3",
     "styled-components": "^5.2.1",
     "typesafe-actions": "^5.1.0"
   },
